fix(Core/Player): Use uint8 for combat rating loop index (#23956)

+++ src/server/game/Entities/Player/PlayerUpdates.cpp
@@ -701,7 +701,7 @@ void Player::UpdateRating(CombatRating cr)
 
 void Player::UpdateAllRatings()
 {
-    for (uint cr = 0; cr < MAX_COMBAT_RATING; ++cr)
+    for (uint8 cr = 0; cr < MAX_COMBAT_RATING; ++cr)
         UpdateRating(CombatRating(cr));
 }
